EXTENSION: Add MVD_PEXT1_EXTRA_PFS.
mvdsv and ezQuake had a broken implementation of FTE_PEXT_TRANS since
inception. To fix this, future versions of ezQuake will enable this
extension in which case mvdsv will write extra playerflags.

To accomodate for these bits PF_ONGROUND and PF_SOLID will be moved if
the extra byte is written. This is how FTE client and server works.